deploy_requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
emergentintegrations
python-multipart>=0.0.9